bitwise

+++ bitwise.go
@@ -1,31 +1,31 @@
 package main
 
 import (
-	"log"
+    "log"
 )
 
 func main() {
-	var a, b, c, d int
-	a = 0x67
-	b = 0x56
-	c = 0x88
-	d = 0x27
+    var a, b, c, d uint32
+    a = 0x67
+    b = 0x56
+    c = 0x88
+    d = 0x27
 
-	x := a | b
-	y := a | b | c
+    x := uint32(a | b)
+    y := a | b | c
 
-	log.Printf("%#x\n", x)
-	log.Printf("%#x\n", y)
+    log.Printf("%#x\n", x)
+    log.Printf("%#x\n", y)
 
-	m := x & a
-	n := x & d
-	log.Printf("%#x\n", m)
-	log.Printf("%#x\n", n)
+    m := x & a
+    n := x & d
+    log.Printf("%#x\n", m)
+    log.Printf("%#x\n", n)
 
-	if (x & a) != 0 {
-		log.Printf("%#x\n", x&a)
-	}
-	if (x & d) != 0 {
-		log.Printf("%#x\n", x&d)
-	}
+    if (x & a) != 0 {
+        log.Printf("%#x\n", x&a)
+    }
+    if (x & d) != 0 {
+        log.Printf("%#x\n", x&d)
+    }
 }
